// Magrathea/src/app/models/config-data.model.ts
export class ConfigData {
    SML: boolean;
    SMR: boolean;
    SMV: boolean;
    SPminF: number;
    SPminI: number;
    SSID_name: string;
    SSID_pass: string;
    SpHum: number;
    SpTemp: number;
    WifiIp: string;
    contLamp: boolean;
    contRieg: boolean;
    contVent: boolean;
    lampara: boolean;
    riego: boolean;
    ventilador: boolean;
}